Rename client/lib to client/components
